Refactor KFP Provider structure
